pyproject.toml
[tool.poetry]
name = "rugby_stats_scraper"
version = "0.1.0"
description = ""
authors = ["jme-taylor <jamiecharlestaylor94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
beautifulsoup4 = "^4.11.1"
pandas = "^1.4.2"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
jupyterlab = "^3.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
